Added GameSenderService
