+++ Modules/GameSender/GameSender.Domain/GameMessages.cs
@@ -7,5 +7,7 @@
         public const string LeaveTheRoom = "Выйти из комнаты";
         public const string GoToPlayers = "Стать играком";
         public const string DenyConnect = "Отказаться";
+
+        public const string ToAnswer = "Ответить";
     }
 }
