+++ Modules/GameSender/GameSender.Infrastructure.Interfaces/IGameSenderService.cs
@@ -11,6 +11,8 @@ public interface IGameSenderService
     // Task<bool> SendConnectedPlayerActions(long playerId);
     Task<bool> SendKickedMessage(long playerId);
     Task<bool> SendLoseGameInFinalRound(long playerId);
+    Task<bool> SendBaseGameButton(long playerId, string text);
+    Task<bool> SendRedButton(long playerId);
 
     Task<Message?> SendMessageAsync(long playerId, string text);
     Task<Message?> ForwardMessageAsync(long playerId, long pictureChatId, long pictureMessageId);
