--- Modules/Games/Game.Domain/Events/Questions/GameIsReadyToReceiveAnswersEvent.cs
@@ -1,9 +1,0 @@
-﻿using Prism.Events;
-
-namespace Game.Domain.Events.Questions
-{
-    /// <summary>
-    /// Игра готова принимать ответы.
-    /// </summary>
-    public class GameIsReadyToReceiveAnswersEvent : PubSubEvent;
-}
