--- Modules/Games/Game.Domain/Events/Questions/QuestionsSendedEvent.cs
@@ -1,9 +1,0 @@
-﻿using Prism.Events;
-
-namespace Game.Domain.Events.Questions
-{
-    /// <summary>
-    /// Вопросы разосланы игрокам.
-    /// </summary>
-    public class QuestionsIsSentEvent : PubSubEvent;
-}
