+++ Modules/Games/Game.Ui/Views/GameControls/Pages/GamePages/GameWinnerViewModel.cs
@@ -3,7 +3,6 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using System.Windows.Input;
 using Common.Core.Prism;
-using Common.Core.Prism.Regions;
 using Common.Core.Views;
 using DataDomain.Rooms;
 using Game.Domain.Data;
