+++ Modules/Games/Game.Ui/Views/GameControls/Pages/GamePages/Topics/AllTopicsNameViewModel.cs
@@ -1,10 +1,8 @@
 ﻿using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
-using System.Reflection.Metadata;
 using System.Windows.Input;
 using Common.Core.Prism;
-using Common.Core.Prism.Regions;
 using Common.Core.Views;
 using DataDomain.Rooms.Rounds;
 using Game.Domain.Data;
